flask_app/requirements.txt
flask
pyjwt
torch
sentence-transformers
transformers